requirements.txt
flask
httpx
gevent
loguru
psutil
pyyaml
pydantic
requests
termcolor
websockets
websocket-client
